Bugfix - Locations not reporting open
Locations were not reporting that they were open when they did have
an hour that covered the current time. This was due to the way that
Location#status was reducing the hours of a location. I also
discovered that having a call to map was completely unnecessary and
reduce was all that was required.